requirements/dev.txt
black
flake8
flake8-black
isort>=5.7
pytest
pytest-cov
torch-pruning
